requirements.txt
voluptuous==0.13.1
PyYAML==6.0